requirements.txt
# Core AI/ML
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.2

# Vector Database
chromadb>=0.5.0
faiss-cpu>=1.7.4

# Backend APIs
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# HTTP & Web
requests>=2.31.0

# Frontend
streamlit>=1.32.0
pillow>=10.0.0

# Utilities
typing-extensions>=4.0.0
numpy>=1.21.0
